Add files via upload
This is a ML based project for predicting the status of patient regarding Parkinson's Disease. This system will provide the results based on available data set. The data of the patient will be added to the system as per given in the CSV file excluding the status row (1 or 0). The Test Accuracy of this model is 87% while the train Accuracy is 88%.